+++ src/core/kv-storage/engines/cookie/engine.ts
@@ -17,11 +17,15 @@ export default class CookieEngine extends StringEngine {
 	}
 
 	protected override set serializedData(value: string) {
+		if (this.date == null) {
+			this.date = Date.now();
+		}
+
 		if (value.length === 0) {
-			this.cookies.remove(this.cookieName, Object.select(this.setOptions, ['path', 'domains']));
+			this.cookies.remove(this.cookieName, Object.select(this.actualOptions, ['path', 'domains']));
 
 		} else {
-			this.cookies.set(this.cookieName, value, this.setOptions);
+			this.cookies.set(this.cookieName, value, this.actualOptions);
 		}
 	}
 
@@ -39,6 +43,37 @@ export default class CookieEngine extends StringEngine {
 	 * Additional options for setting cookies
 	 */
 	protected setOptions: SetOptions;
+
+	/**
+	 * The date of setting cookie
+	 */
+	protected date: CanUndef<number>;
+
+	/**
+	 * Actual options for setting cookies.
+	 * Ensure cookie options, especially `maxAge`, are kept up-to-date.
+	 *
+	 * @example
+	 * ```js
+	 * const store = new CookieEngine('bla', {maxAge: 600});
+	 * store.set('foo', 5); // cookie is added with max-age=600
+	 * ... // after 60s
+	 * store.set('bar', 10); // cookie is added with max-age=540
+	 * ```
+	 */
+	protected get actualOptions(): SetOptions {
+		if (this.setOptions.maxAge == null) {
+			return this.setOptions;
+		}
+
+		const
+			passedTime = Math.floor((Date.now() - this.date!) / 1000);
+
+		return {
+			...this.setOptions,
+			maxAge: this.setOptions.maxAge - passedTime
+		};
+	}
 
 	/**
 	 * @param cookieName - the name of the cookie in which the data is stored