+++ CHANGELOG.md
@@ -11,6 +11,16 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v4.0.0-beta.119 (2024-08-02)
+
+#### :rocket: New Feature
+
+* Added decorator for cookie store `core/cookies`
+
+#### :bug: Bug Fix
+
+* Fixed a bug of re-applying `maxAge` parameter on storage update `core/kv-storage/engines/cookie`
+
 ## v4.0.0-beta.118 (2024-08-01)
 
 #### :bug: Bug Fix